@@ -7,78 +7,77 @@
   ],
   "license": "MIT",
   "scripts": {
-    "start": "yarn tokens && start-storybook -s .storybook/public -p 6006",
+    "start": "yarn tokens && storybook dev -p .storybook/public -p 6006",
     "build": "webpack --mode=production",
-    "storybook:build": "build-storybook -s .storybook/public",
+    "storybook build": "storybook build -o dist/storybook --loglevel verbose",
     "storybook:deploy": "gh-pages -d storybook-static",
     "add-component": "node .add-component/index.js",
     "tokens": "style-dictionary build --config ./src/tokens/config.js",
-    "test": "yarn creevey",
-    "test:update": "yarn creevey --update",
-    "test:ui": "yarn creevey --ui",
     "semantic-release": "semantic-release"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.8",
-    "@babel/preset-env": "^7.13.9",
-    "@babel/preset-react": "^7.12.13",
-    "@mariusmarais/tailwind-cascade": "^0.0.4",
-    "@semantic-release/changelog": "^5.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@storybook/addon-essentials": "6.3.7",
+    "@babel/core": "^7.22.8",
+    "@babel/preset-env": "^7.22.7",
+    "@babel/preset-react": "^7.22.5",
+    "@semantic-release/changelog": "^6.0.3",
+    "@semantic-release/git": "^10.0.1",
+    "@storybook/addon-essentials": "7.0.26",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/react": "6.3.7",
-    "@storybook/storybook-deployer": "^2.8.10",
-    "@storybook/theming": "^6.3.7",
-    "@svgr/webpack": "^5.5.0",
+    "@storybook/blocks": "^7.1.0",
+    "@storybook/react-webpack5": "^7.1.0",
+    "@storybook/testing-library": "^0.2.0",
+    "@storybook/theming": "^7.0.26",
+    "@svgr/webpack": "^8.0.1",
+    "autoprefixer": "^10.4.14",
     "babel-eslint": "^10.1.0",
     "babel-loader": "8.1.0",
-    "babel-plugin-module-resolver": "^4.1.0",
-    "babel-plugin-styled-components": "^1.12.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "conventional-changelog-conventionalcommits": "^4.6.0",
-    "copy-webpack-plugin": "^4.6.0",
-    "creevey": "^0.7.27",
-    "css-loader": "^0.28.11",
-    "eslint": "^7.26.0",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-mdx": "^1.14.1",
-    "eslint-plugin-react": "^7.22.0",
-    "file-loader": "^1.1.11",
-    "gh-pages": "^3.2.0",
-    "img-loader": "^2.0.1",
+    "babel-plugin-module-resolver": "^5.0.0",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.8.1",
+    "eslint": "^8.44.0",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-mdx": "^2.1.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-storybook": "^0.6.12",
+    "file-loader": "^6.2.0",
+    "gh-pages": "^5.0.0",
     "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "^1.6.0",
-    "postcss": "^8.2.15",
-    "postcss-loader": "^4.2.0",
-    "prettier": "^2.2.1",
-    "semantic-release": "^19.0.5",
-    "storybook-addon-designs": "^6.0.1",
-    "style-dictionary": "^2.10.3",
-    "url-loader": "^0.6.2",
-    "webpack": "4.44.2",
-    "webpack-cli": "^4.5.0"
+    "@carbon/icons-react": "^11.22.1",
+    "mini-css-extract-plugin": "^2.7.6",
+    "prettier": "^3.0.0",
+    "semantic-release": "^21.0.7",
+    "storybook": "^7.1.0",
+    "style-dictionary": "^3.8.0",
+    "tailwindcss": "^3.3.3",
+    "url-loader": "^4.1.1",
+    "webpack": "5.88.1",
+    "webpack-cli": "^5.1.4",
+    "react": "17",
+    "react-dom": "17"
   },
   "dependencies": {
-    "@carbon/icons-react": "^11.13.0",
-    "@tailwindcss/typography": "^0.4.1",
-    "classnames": "^2.3.1",
-    "core-js": "^3.9.1",
-    "fast-cartesian": "5.1",
-    "prop-types": "^15.7.2",
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "swiper": "^6.7.5",
-    "tailwindcss": "^2.1.2",
-    "use-debounce": "^7.0.0"
+    "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
+    "@tailwindcss/typography": "^0.5.9",
+    "chokidar": "^3.5.3",
+    "classnames": "^2.3.2",
+    "core-js": "^3.31.1",
+    "fast-cartesian": "8.0.0",
+    "prop-types": "^15.5.7",
+    "swiper": "^10.0.4",
+    "postcss": "^8.4.27",
+    "use-debounce": "^9.0.4"
   },
   "peerDependencies": {
-    "@carbon/icons-react": "^10.32.0",
     "@tailwindcss/typography": "^0.4.1",
     "swiper": "^6.7.5"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/bridge-design/design-system.git"
+  },
+  "eslintConfig": {
+    "extends": [
+      "plugin:storybook/recommended"
+    ]
   }
 }
