@@ -1,5 +1,5 @@
-import { addons } from "@storybook/addons";
-import bridgeTheme from "./bridge.theme";
+import { addons } from "@storybook/manager-api";
+import bridgeTheme from "./bridge";
 
 addons.setConfig({
   theme: bridgeTheme,
