@@ -1,8 +1,6 @@
 module.exports = {
-  plugins: [
-    require("tailwindcss"),
-    require("autoprefixer")({
-      grid: false,
-    }),
-  ],
-};
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  }
+}
