@@ -2,43 +2,36 @@
 /*
  * This webpack config is used for production build of components library
  */
-const webpack = require("webpack");
-const nodeExternals = require("webpack-node-externals");
+const React = require("react");
 const path = require("path");
+const webpack = require("webpack");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
 const paths = {
   libSrc: path.resolve(__dirname, "src"),
   libIndex: [path.resolve(__dirname, "src/index.js")],
   libOutputDir: path.resolve(__dirname, "dist"),
-  libModules: path.resolve(__dirname, "../node_modules"),
 };
 
 module.exports = (_, argv) => {
-  const exts = [
-    nodeExternals({
-      allowlist: /debounce/,
-    }),
-  ];
-
   const env = argv.mode;
-
   process.env.BABEL_ENV = env;
-  process.env.NODE_ENV = env;
 
   let libraryName = "design-system";
   let plugins = [
     new webpack.DefinePlugin({
-      "process.env": {
-        NODE_ENV: JSON.stringify(process.env.NODE_ENV),
-      },
+      "process.env.NODE_ENV": JSON.stringify(env),
     }),
-    new MiniCssExtractPlugin({
-      filename: "styles.css",
-      chunkFilename: "styles.css",
-    }),
-    new webpack.NamedModulesPlugin(),
   ];
+
+  if (env === "production") {
+    plugins.push(
+      new MiniCssExtractPlugin({
+        filename: "styles.css",
+        chunkFilename: "styles.css",
+      })
+    );
+  }
 
   return {
     entry: paths.libIndex,
@@ -46,14 +39,21 @@ module.exports = (_, argv) => {
     output: {
       path: paths.libOutputDir,
       filename: libraryName + ".js",
-      libraryTarget: "commonjs",
+      libraryTarget: "commonjs2",
+      library: {
+        type: "commonjs2",
+      },
     },
     module: {
       rules: [
         {
           test: /\.css$/i,
-          /* Don"t use style-loader if you will impport library into Next.js */
-          use: [MiniCssExtractPlugin.loader, "css-loader"],
+          use: [
+            env === "production"
+              ? MiniCssExtractPlugin.loader
+              : "style-loader", // Use style-loader in development for hot-reloading
+            "css-loader",
+          ],
         },
         {
           test: /\.(js|jsx)$/,
@@ -61,50 +61,54 @@ module.exports = (_, argv) => {
           use: ["babel-loader"],
         },
         {
-          // Apply rule for fonts files
-          test: /\.(woff|woff2|ttf|otf|eot)$/,
-          use: [
-            {
-              // Using file-loader too
-              loader: "file-loader",
-              options: {
-                outputPath: "fonts",
-                publicPath: "fonts",
-              },
-            },
-          ],
+          test: /\.(jpe?g|png|gif)$/i,
+          type: "asset",
+          generator: {
+            filename: "images/[name][ext]",
+          },
         },
         {
-          test: /\.(jpe?g|png|gif)$/i,
-          use: ["url-loader?limit=10000", "img-loader"],
+          // Apply rule for fonts files
+          test: /\.(woff|woff2|ttf|otf|eot)$/,
         },
         {
           test: /\.svg$/,
-
-          issuer: {
-            test: /\.(js|ts)x?$/,
-          },
+          include: [path.resolve(__dirname, "src")],
           use: [
             {
               loader: "@svgr/webpack",
               options: {
-                svgoConfig: {
-                  plugins: {
-                    removeViewBox: false,
-                  },
-                },
+                svgo: false,
               },
             },
-            "file-loader",
+            {
+              loader: "url-loader",
+              options: {
+                limit: 8192,
+                outputPath: "images",
+                publicPath: "images",
+                fallback: require.resolve("file-loader"),
+              },
+            },
           ],
         },
       ],
     },
-    externals: exts,
     plugins: plugins,
-    mode: "production",
-    node: {
-      fs: "empty",
+    mode: env,
+    externals: {
+      react: "commonjs react",
+      "react-dom": "commonjs react-dom",
+      debounce: "debounce", // Assuming debounce is a package name to be externalized
+    },
+    resolve: {
+      extensions: [".js", ".jsx"],
+    },
+    performance: {
+      hints: false,
+    },
+    optimization: {
+      minimize: true,
     },
   };
 };
